+++ lace/include/lace/codegen/LIRCodegen.hpp
@@ -22,6 +22,7 @@
 #include "lir/graph/Builder.hpp"
 #include "lir/graph/CFG.hpp"
 #include "lir/graph/Function.hpp"
+#include "lir/graph/Type.hpp"
 #include "lir/machine/Machine.hpp"
 
 namespace lace {
@@ -46,6 +47,9 @@ public:
 private:
     /// Lower the given lace |type| to its LIR equivelant, where possible.
     lir::Type* to_lir_type(const QualType& type);
+
+    lir::Function* get_intrinsic(const std::string& name, lir::Type* result, 
+                                 const std::vector<lir::Type*>& args = {});
 
     /// Attempt to inject a boolean comparison unto the given |value|, such
     /// that the result is some form of comparison of a boolean type.