+++ lace/source/codegen/LIRCodegenStmt.cpp
@@ -6,6 +6,7 @@
 #include "lace/codegen/LIRCodegen.hpp"
 #include "lace/tree/Defn.hpp"
 #include "lace/tree/Stmt.hpp"
+#include "lir/graph/Type.hpp"
 
 using namespace lace;
 
@@ -142,14 +143,48 @@ void LIRCodegen::codegen_stop(const StopStmt* stmt)  {
 }
 
 void LIRCodegen::codegen_return(const RetStmt* stmt) {
-    lir::Value* value = nullptr;
-
-    if (stmt->has_expr()) {
-        value = codegen_valued_expression(stmt->get_expr());
-        assert(value);
+    if (!stmt->has_expr()) {
+        m_builder.build_ret();
+        return;
     }
 
-    m_builder.build_ret(value);
+    lir::Type* type = to_lir_type(stmt->get_expr()->get_type());
+    
+    if (m_mach.is_scalar(type)) {
+        lir::Value* value = codegen_valued_expression(stmt->get_expr());
+        assert(value);
+
+        m_builder.build_ret(value);
+    } else {
+        // Return type must be non-scalar/aggregate, so we write changes to the
+        // aret argument.
+        lir::Value* value = codegen_addressed_expression(stmt->get_expr());
+        assert(value);
+
+        lir::Function* copy = get_intrinsic(
+            "__copy", 
+            lir::VoidType::get(m_cfg), 
+            { 
+                lir::PointerType::get_void_pointer(m_cfg),
+                lir::PointerType::get_void_pointer(m_cfg),
+                lir::IntegerType::get_i64_type(m_cfg),
+            }
+        );
+
+        lir::Local* aretL = m_func->get_local(".ret");
+        assert(aretL);
+
+        lir::Value* aretP = m_builder.build_load(
+            lir::PointerType::get(m_cfg, type), aretL);
+
+        m_builder.build_call(copy->get_type(), copy, {
+            aretP,
+            value,
+            lir::Integer::get(m_cfg, lir::IntegerType::get_i64_type(m_cfg), m_mach.get_size(type))
+        });
+
+        m_builder.build_ret();
+    }
 }
 
 void LIRCodegen::codegen_rune_statement(const RuneStmt* stmt) {